Refactor and cleanup
- took out spinner (make sense to just stick with the auto complete view)
- auto complete view is now part of a new class DemoSceneManager which provides an interface to
loadScene which is implemented in the MainActvity
- Use more-label variants from cdn
- autocompletetextview highlights the entire text on getting focus
