+++ platforms/android/demo/build.gradle
@@ -25,5 +25,5 @@ android {
 
 dependencies {
   compile project(path: ':tangram')
-    compile 'com.android.support:design:25.3.1'
+  compile 'com.android.support:design:25.3.1'
 }
