--- platforms/android/demo/src/main/res/layout/simple_spinner_dropdown_item.xml
@@ -1,8 +1,0 @@
-<CheckedTextView xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@android:id/text1"
-    style="?android:attr/spinnerDropDownItemStyle"
-    android:layout_width="match_parent"
-    android:layout_height="44dp"
-    android:ellipsize="marquee"
-    android:singleLine="true"
-    android:textColor="@android:color/background_light" />
